+++ src/test/java/de/codescape/jira/plugins/scrumpoker/ScrumPokerTestDatabaseUpdater.java
@@ -1,0 +1,24 @@
+package de.codescape.jira.plugins.scrumpoker;
+
+import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerSession;
+import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerSetting;
+import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerVote;
+import net.java.ao.EntityManager;
+import net.java.ao.test.jdbc.DatabaseUpdater;
+
+/**
+ * Database updater used for testing Active Object implementation.
+ */
+public class ScrumPokerTestDatabaseUpdater implements DatabaseUpdater {
+
+    @Override
+    public void update(EntityManager entityManager) throws Exception {
+        //noinspection unchecked
+        entityManager.migrate(
+            ScrumPokerSession.class,
+            ScrumPokerSetting.class,
+            ScrumPokerVote.class
+        );
+    }
+
+}