+++ changelog.md
@@ -8,6 +8,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * feat: make Scrum Poker accessible from Jira Software boards menu
 * docs: update documentation for joining an active session
 * refactor: use links starting with uppercase for Scrum Poker
+* feat: move Scrum Poker configuration into database (#31)
 
 ## [3.4.0] - 2018-09-15